requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
uvloop==0.19.0
python-socketio==5.11.2
python-multipart==0.0.9
SQLAlchemy==2.0.30
aiosqlite==0.20.0
asyncpg==0.29.0
redis==5.0.4
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.1
Pillow==10.3.0
Jinja2==3.1.4
